+++ README.md
@@ -1,5 +1,6 @@
 ### 以下の資料を元に作成しました。
 - [【SSR、SSG、ISRでブログを作る】Nextjs14初心者入門](https://zenn.dev/y_ta/books/eec3b78567aeeb)
+- [Nextjs14初心者入門-Loading UIとStreaming](https://zenn.dev/y_ta/articles/b1908ec6af6819)
 
 ### デモ
 - [デモページ](http://zenn-next-demo.vercel.app)
